+++ src/api/counted_list.ts
@@ -73,6 +73,11 @@ export default class CountedList<T extends NamedTootCount> {
         return new CountedList<T>(this.objs.filter(predicate), this.source);
     }
 
+    /** Standard Array.forEach() approximation that invokes a callback for each object in the objs array. */
+    forEach(callback: (obj: T, i?: number) => void): void {
+        this.objs.forEach((obj, i) => callback(obj, i));
+    }
+
     /**
      * Returns the object in the list with the given name (case-insensitive) if it exists.
      * @param {string} name - The name of the object to retrieve.